+++ python/tvm/relax/backend/adreno/__init__.py
@@ -18,6 +18,7 @@
 from .pipeline import (
     finalize_passes,
     get_default_pipeline,
+    dataflow_lower_passes,
     legalize_passes,
     library_dispatch_passes,
 )
