+++ tests/python/relax/backend/clml/mod_utils.py
@@ -29,10 +29,8 @@ from tvm.script.ir_builder import IRBuilder
 from tvm.script.ir_builder import relax as relax_builder
 from tvm.relax.backend.adreno import clml
 
-from utils import build_and_run, run_compare, verify_codegen
 
-
-def get_relax_conv2d_module(
+def get_relax_conv2d_mod(
     data_shape,
     weight_shape,
     stride,
@@ -206,100 +204,7 @@ def get_clml_conv2d_codegen(
     return nodes
 
 
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "kernel_h, kernel_w, padding, stride, dilation, out_channels, shape, has_bias, has_bn, has_activation, has_pad, is_depthwise",
-    [
-        (3, 3, (1, 1), (1, 1), (1, 1), 64, (3, 224, 224), False, True, False, True, False),
-        (3, 3, (1, 1), (1, 1), (1, 1), 64, (3, 224, 224), False, True, False, False, False),
-        (5, 5, (2, 2), (1, 1), (1, 1), 16, (16, 64, 64), False, True, True, False, False),
-        (7, 7, (3, 3), (2, 2), (1, 1), 32, (3, 224, 224), True, False, True, True, False),
-        (3, 3, (0, 0), (1, 1), (1, 1), 512, (256, 14, 14), True, False, True, False, False),
-        (1, 1, (0, 0), (1, 1), (1, 1), 1024, (512, 7, 7), True, False, True, False, False),
-        (1, 3, (0, 0), (1, 1), (1, 1), 64, (64, 7, 7), True, False, True, False, False),
-        (3, 1, (0, 0), (1, 1), (1, 1), 64, (64, 7, 7), False, True, True, True, False),
-    ],
-)
-def test_conv2d_offload(
-    kernel_h,
-    kernel_w,
-    padding,
-    stride,
-    dilation,
-    out_channels,
-    shape,
-    has_bias,
-    has_bn,
-    has_activation,
-    has_pad,
-    is_depthwise,
-    dtype,
-    rpc,
-):
-    low, high = 0, 1
-    data_shape = (1, *shape)
-    if is_depthwise:
-        groups = data_shape[1] // out_channels
-    else:
-        groups = 1
-    padding = (padding[0], padding[1], padding[0], padding[1])
-
-    weight_format = "IOHW" if is_depthwise else "OIHW"
-    weight_shape = (out_channels, data_shape[1] // groups, kernel_h, kernel_w)
-
-    data = np.random.uniform(low, high, size=data_shape).astype(dtype)
-    weight = np.random.uniform(low, high, size=weight_shape).astype(dtype)
-    bias = np.random.uniform(low, high, size=(1, weight_shape[0], 1, 1)).astype(dtype)
-
-    gamma = np.random.uniform(low, high, size=(weight_shape[0],)).astype(dtype)
-    beta = np.random.uniform(low, high, size=(weight_shape[0],)).astype(dtype)
-    mean = np.random.uniform(low, high, size=(weight_shape[0],)).astype(dtype)
-    variance = np.random.uniform(low, high, size=(weight_shape[0],)).astype(dtype)
-
-    inputs = [data]
-    params_np = {"weight": weight}
-    if has_bias:
-        params_np["bias"] = bias
-    if has_bn:
-        params_np.update({"gamma": gamma, "beta": beta, "mean": mean, "variance": variance})
-
-    mod = get_relax_conv2d_module(
-        data_shape,
-        weight_shape,
-        stride=stride,
-        dilation=dilation,
-        padding=padding,
-        weight_layout=weight_format,
-        groups=groups,
-        dtype=dtype,
-        has_bias=has_bias,
-        has_bn=has_bn,
-        has_activation=has_activation,
-        has_pad=has_pad,
-        is_depthwise=is_depthwise,
-    )
-
-    clml_codegen = get_clml_conv2d_codegen(
-        data_shape,
-        weight_shape,
-        stride=stride,
-        dilation=dilation,
-        padding=padding,
-        weight_layout=weight_format,
-        groups=groups,
-        dtype=dtype,
-        has_bias=has_bias,
-        has_bn=has_bn,
-        has_activation=has_activation,
-        has_pad=has_pad,
-        is_depthwise=is_depthwise,
-    )
-
-    run_compare(mod, inputs, params_np, rpc, clml_codegen)
-
-
-def get_relax_conv2d_transpose_module(
+def get_relax_conv2d_transpose_mod(
     data_shape,
     weight_shape,
     channels,
@@ -333,7 +238,7 @@ def get_relax_conv2d_transpose_module(
     return tvm.IRModule({"main": func})
 
 
-def _get_conv2d_transpose_expected_codegen(
+def get_conv2d_transpose_expected_codegen(
     dshape, kshape, channels, kernel_size, strides, padding, dilation, dtype, output_shape
 ):
     attrs = {
@@ -374,295 +279,70 @@ def _get_conv2d_transpose_expected_codegen(
     return exp_codegen
 
 
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "dshape, kshape, channels, kernel_size, strides, padding, out_shape",
-    [
-        ((1, 256, 100, 100), (64, 256, 4, 4), 64, (4, 4), (2, 2), (0, 0, 0, 0), (1, 64, 202, 202)),
-        ((1, 64, 200, 200), (64, 64, 4, 4), 64, (4, 4), (2, 2), (1, 1, 1, 1), (1, 64, 400, 400)),
-        ((1, 64, 200, 200), (64, 64, 4, 4), 64, (4, 4), (2, 2), (1, 1, 1, 1), (1, 64, 400, 400)),
-        ((1, 64, 400, 400), (16, 64, 4, 4), 16, (4, 4), (2, 2), (1, 1, 1, 1), (1, 16, 800, 800)),
-    ],
-)
-def test_conv2d_transpose(
-    dshape, kshape, channels, kernel_size, strides, padding, rpc, dtype, out_shape
-):
-    low, high = -1, 1
-    data = np.random.uniform(low, high, size=dshape).astype(dtype)
-    weight = np.random.uniform(low, high, size=kshape).astype(dtype)
+def get_batchnorm_mod(data_shape, channels, axis, epsilon, dtype):
+    with IRBuilder() as builder:
+        with relax_builder.function():
+            R.func_name("main")
+            data = R.arg("data", R.Tensor(data_shape, dtype))
+            gamma = R.arg("gamma", R.Tensor((channels,), dtype))
+            beta = R.arg("beta", R.Tensor((channels,), dtype))
+            mean = R.arg("moving_mean", R.Tensor((channels,), dtype))
+            variance = R.arg("moving_var", R.Tensor((channels,), dtype))
+            with R.dataflow() as frame:
+                output = R.emit(
+                    R.nn.batch_norm(data, gamma, beta, mean, variance, axis, epsilon)[0]
+                )
+                R.output(output)
 
-    inputs = [data]
-    params_np = {"weight": weight}
-
-    mod = get_relax_conv2d_transpose_module(
-        dshape,
-        kshape,
-        channels=channels,
-        stride=strides,
-        padding=padding,
-        dtype=dtype,
-    )
-
-    exp_codegen = _get_conv2d_transpose_expected_codegen(
-        dshape=dshape,
-        kshape=kshape,
-        channels=channels,
-        kernel_size=kernel_size,
-        strides=strides,
-        padding=padding,
-        dilation=(1, 1),
-        dtype=dtype,
-        output_shape=out_shape,
-    )
-
-    mod = tvm.relax.transform.BindParams("main", params_np)(mod)
-    clml_mod = clml.OpenCLMLOffLoad()(mod)
-    verify_codegen(clml_mod, exp_codegen)
-
-
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "trials",
-    [
-        [(1, 64, 14, 14), 1, 3e-4],
-        [(1, 14, 256, 256), 1, 3e-4],
-        [(1, 14, 256, 256), 1, 3e-4],
-        [(1, 256, 1, 1), 1, 3e-4],
-    ],
-)
-def test_batchnorm(dtype, trials, rpc):
-    low, high = 0, 1
-    if clml.clml_sdk_version() < 3:
-        print("Skip due to unsupported CLML version:", clml.clml_sdk_version())
-        return
-
-    (input_shape, axis, epsilon) = trials
-    channels = input_shape[axis]
-
-    def _get_batchnorm_model(data_shape, channels, axis, epsilon, dtype):
-
-        with IRBuilder() as builder:
-            with relax_builder.function():
-                R.func_name("main")
-                data = R.arg("data", R.Tensor(data_shape, dtype))
-                gamma = R.arg("gamma", R.Tensor((channels,), dtype))
-                beta = R.arg("beta", R.Tensor((channels,), dtype))
-                mean = R.arg("moving_mean", R.Tensor((channels,), dtype))
-                variance = R.arg("moving_var", R.Tensor((channels,), dtype))
-                with R.dataflow() as frame:
-                    output = R.emit(
-                        R.nn.batch_norm(data, gamma, beta, mean, variance, axis, epsilon)[0]
-                    )
-                    R.output(output)
-
-                R.func_ret_value(frame.output_vars[0])
-
-            func = builder.get()
-            return tvm.IRModule({"main": func})
-
-    def _get_axis_tuple(axis):
-        if axis == 0:
-            return (1, 2, 3)
-        elif axis == 1:
-            return (0, 2, 3)
-        elif axis == 2:
-            return (0, 1, 3)
-        else:
-            return (0, 1, 2)
-
-    data = np.random.uniform(low, high, size=(input_shape)).astype(dtype)
-    gamma = np.random.uniform(low, high, size=(channels)).astype(dtype)
-    beta = np.random.uniform(low, high, size=(channels)).astype(dtype)
-    mean = np.mean(data, _get_axis_tuple(axis), keepdims=False)
-    variance = np.var(data, _get_axis_tuple(axis), keepdims=False)
-
-    inputs = [data]
-    params_np = {"gamma": gamma, "beta": beta, "moving_mean": mean, "moving_var": variance}
-    mod = _get_batchnorm_model(input_shape, channels, axis, epsilon, dtype)
-    exp_codegen = [
-        {
-            "attrs": {"dtype": [[dtype]], "shape": [[input_shape]]},
-            "name": "",
-            "op": "input",
-        },
-        {"attrs": {"dtype": [[dtype]], "shape": [[[channels]]]}, "name": "", "op": "const"},
-        {"attrs": {"dtype": [[dtype]], "shape": [[[channels]]]}, "name": "", "op": "const"},
-        {"attrs": {"dtype": [[dtype]], "shape": [[[channels]]]}, "name": "", "op": "const"},
-        {"attrs": {"dtype": [[dtype]], "shape": [[[channels]]]}, "name": "", "op": "const"},
-        {
-            "attrs": {
-                "axis": [[str(axis)]],
-                "center": [["1"]],
-                "dtype": [[dtype]],
-                "clml_version": [["3"]],
-                "momentum": [["0.10000000000000001"]],
-                "epsilon": [["0.00029999999999999997"]],
-                "num_inputs": "5",
-                "num_outputs": "1",
-                "scale": [["1"]],
-                "shape": [[input_shape]],
-            },
-            "inputs": [[0, 0, 0], [1, 0, 0], [2, 0, 0], [3, 0, 0], [4, 0, 0]],
-            "name": "",
-            "op": "kernel",
-        },
-    ]
-    run_compare(mod, inputs, params_np, rpc, exp_codegen)
-
-
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "a_shape, b_shape, op",
-    [
-        ((1, 64, 14, 14), (1, 64, 14, 14), R.add),
-        ((1, 256), (1, 256), R.add),
-        ((1, 64, 14, 14), (1, 64, 14, 14), R.subtract),
-        ((1, 256), (1, 256), R.subtract),
-        ((1, 64, 14, 14), (1, 64, 14, 14), R.multiply),
-        ((1, 256), (1, 256), R.multiply),
-        ((1, 64, 14, 14), (1, 64, 14, 14), R.divide),
-        ((1, 256), (1, 256), R.divide),
-        ((1, 64, 14, 14), (1, 64, 14, 14), R.minimum),
-        ((1, 256), (1, 256), R.minimum),
-        ((1, 64, 14, 14), (1, 64, 14, 14), R.maximum),
-        ((1, 256), (1, 256), R.maximum),
-    ],
-)
-@tvm.testing.requires_openclml
-def test_binary_ops(a_shape, b_shape, op, rpc, dtype):
-    low, high = 0, 1
-
-    def _get_model(a_shape, b_shape, op):
-        with IRBuilder() as builder:
-            with relax_builder.function():
-                R.func_name("main")
-                a = R.arg("a", R.Tensor(a_shape, dtype))
-                b = R.arg("b", R.Tensor(b_shape, dtype))
-
-                with R.dataflow() as frame:
-                    output = R.emit(op(a, b))
-                    R.output(output)
-
-                R.func_ret_value(frame.output_vars[0])
+            R.func_ret_value(frame.output_vars[0])
 
         func = builder.get()
-
-        a_data = np.random.uniform(low, high, size=(a_shape)).astype(dtype)
-        b_data = np.random.uniform(low, high, size=(b_shape)).astype(dtype)
-
-        return (tvm.IRModule({"main": func}), (a_data, b_data))
-
-    def _verify(mod, inputs):
-        expected_codegen_str = [
-            {
-                "attrs": {
-                    "dtype": [[dtype]],
-                    "shape": [[a_shape]],
-                },
-                "name": "",
-                "op": "input",
-            },
-            {
-                "attrs": {
-                    "dtype": [[dtype]],
-                    "shape": [[b_shape]],
-                },
-                "name": "",
-                "op": "input",
-            },
-            {
-                "attrs": {
-                    "clml_version": [["3"]],
-                    "dtype": [[dtype]],
-                    "num_inputs": "2",
-                    "num_outputs": "1",
-                    "shape": [[a_shape]],
-                },
-                "inputs": [[0, 0, 0], [1, 0, 0]],
-                "name": "",
-                "op": "kernel",
-            },
-        ]
-        run_compare(mod, inputs, {}, rpc, expected_codegen_str)
-
-    (mod, inputs) = _get_model(a_shape, b_shape, op)
-
-    _verify(mod, inputs)
+        return tvm.IRModule({"main": func})
 
 
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize(
-    "dtype",
-    [
-        "float32",
-    ],
-)
-@pytest.mark.parametrize(
-    "a_shape, op",
-    [
-        ((1, 64, 14, 14), R.nn.relu),
-        ((1, 256, 1, 1), R.nn.relu),
-        ((1, 14, 256, 256), R.nn.relu),
-        ((1, 14, 14, 256), R.nn.relu),
-    ],
-)
-@tvm.testing.requires_openclml
-def test_unary_ops(a_shape, op, rpc, dtype):
+def get_binary_op_mod(a_shape, b_shape, op, dtype):
+    with IRBuilder() as builder:
+        with relax_builder.function():
+            R.func_name("main")
+            a = R.arg("a", R.Tensor(a_shape, dtype))
+            b = R.arg("b", R.Tensor(b_shape, dtype))
+
+            with R.dataflow() as frame:
+                output = R.emit(op(a, b))
+                R.output(output)
+
+            R.func_ret_value(frame.output_vars[0])
+
+    func = builder.get()
+
     low, high = 0, 1
+    a_data = np.random.uniform(low, high, size=(a_shape)).astype(dtype)
+    b_data = np.random.uniform(low, high, size=(b_shape)).astype(dtype)
 
-    def _get_model(a_shape, op):
-        with IRBuilder() as builder:
-            with relax_builder.function():
-                R.func_name("main")
-                a = R.arg("a", R.Tensor(a_shape, dtype))
-
-                with R.dataflow() as frame:
-                    output = R.emit(op(a))
-                    R.output(output)
-
-                R.func_ret_value(frame.output_vars[0])
-
-        func = builder.get()
-
-        a_data = np.random.uniform(low, high, size=(a_shape)).astype(dtype)
-
-        return (tvm.IRModule({"main": func}), (a_data,))
-
-    def _verify(mod, inputs):
-        expected_codegen_str = [
-            {
-                "attrs": {
-                    "dtype": [[dtype]],
-                    "shape": [[a_shape]],
-                },
-                "name": "",
-                "op": "input",
-            },
-            {
-                "attrs": {
-                    "activation_type": [["relu"]],
-                    "clml_version": [["3"]],
-                    "dtype": [[dtype]],
-                    "num_inputs": "1",
-                    "num_outputs": "1",
-                    "shape": [[a_shape]],
-                },
-                "inputs": [[0, 0, 0]],
-                "name": "",
-                "op": "kernel",
-            },
-        ]
-        run_compare(mod, inputs, {}, rpc, expected_codegen_str)
-
-    (mod, inputs) = _get_model(a_shape, op)
-
-    _verify(mod, inputs)
+    return (tvm.IRModule({"main": func}), (a_data, b_data))
 
 
-def get_relax_maxpool_module(
+def get_unary_op_mod(a_shape, op, dtype):
+    with IRBuilder() as builder:
+        with relax_builder.function():
+            R.func_name("main")
+            a = R.arg("a", R.Tensor(a_shape, dtype))
+
+            with R.dataflow() as frame:
+                output = R.emit(op(a))
+                R.output(output)
+
+            R.func_ret_value(frame.output_vars[0])
+
+    func = builder.get()
+
+    low, high = 0, 1
+    a_data = np.random.uniform(low, high, size=(a_shape)).astype(dtype)
+
+    return (tvm.IRModule({"main": func}), (a_data,))
+
+
+def get_relax_maxpool_mod(
     data_shape, dtype, pool_size, stride=None, dilation=(1, 1), padding=(0, 0), has_pad=False
 ):
     """
@@ -756,37 +436,7 @@ def get_maxpool_expected_codegen(input_shape, pool_size, stride, padding, pool_t
     return exp_codegen
 
 
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "trials",
-    [
-        [(1, 64, 147, 147), (3, 3), (2, 2), (1, 1), (0, 0, 0, 0), False],
-        [(1, 256, 17, 17), (3, 3), (1, 1), (1, 1), (0, 0, 0, 0), False],
-        [(1, 1024, 14, 14), (3, 3), (1, 1), (1, 1), (0, 0, 0, 0), False],
-        [(1, 32, 256, 256), (3, 3), (2, 2), (1, 1), (1, 1, 1, 1), True],
-        [(1, 32, 256, 256), (3, 3), (2, 2), (1, 1), (0, 1, 0, 1), True],
-        [(1, 32, 256, 256), (2, 2), (2, 2), (1, 1), (1, 1, 1, 1), True],
-        [(1, 32, 256, 256), (2, 2), (2, 2), (1, 1), (1, 0, 1, 0), True],
-    ],
-)
-def test_max_pool(dtype, trials, rpc):
-    low, high = -1, 1
-    (input_shape, pool_size, stride, dilation, padding, has_pad) = trials
-    data = np.random.uniform(low, high, size=input_shape).astype(dtype)
-    inputs = [data]
-    mod = get_relax_maxpool_module(
-        input_shape, dtype, pool_size, stride, dilation, padding, has_pad
-    )
-    params_np = {}
-
-    expected_codegen_str = get_maxpool_expected_codegen(
-        input_shape, pool_size, stride, padding, "maxpool2d", dtype
-    )
-    run_compare(mod, inputs, params_np, rpc, expected_codegen_str)
-
-
-def get_relax_avgpool_module(data_shape, dtype, pool_size, stride, dilation, padding, has_pad):
+def get_relax_avgpool_mod(data_shape, dtype, pool_size, stride, dilation, padding, has_pad):
     """
     Args:
         data_shape (tuple): Input tensor shape
@@ -879,36 +529,7 @@ def get_avgpool_expected_codegen(input_shape, pool_size, stride, padding, pool_t
     return exp_codegen
 
 
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "trials",
-    [
-        [(1, 64, 147, 147), (3, 3), (2, 2), (1, 1), (0, 0, 0, 0), False],
-        [(1, 256, 17, 17), (3, 3), (1, 1), (1, 1), (0, 0, 0, 0), False],
-        [(1, 1024, 14, 14), (3, 3), (1, 1), (1, 1), (0, 0, 0, 0), False],
-        [(1, 32, 256, 256), (3, 3), (2, 2), (1, 1), (1, 1, 1, 1), True],
-        [(1, 32, 256, 256), (3, 3), (2, 2), (1, 1), (0, 1, 0, 1), True],
-        [(1, 32, 256, 256), (2, 2), (2, 2), (1, 1), (1, 1, 1, 1), True],
-        [(1, 32, 256, 256), (2, 2), (2, 2), (1, 1), (1, 0, 1, 0), True],
-    ],
-)
-def test_avg_pool(dtype, trials, rpc):
-    low, high = -1, 1
-    (input_shape, pool_size, stride, dilation, padding, has_pad) = trials
-    data = np.random.uniform(low, high, size=input_shape).astype(dtype)
-    inputs = [data]
-    mod = get_relax_avgpool_module(
-        input_shape, dtype, pool_size, stride, dilation, padding, has_pad
-    )
-    params_np = {}
-    exp_codegen_str = get_avgpool_expected_codegen(
-        input_shape, pool_size, stride, padding, "avg_pool2d", dtype
-    )
-    run_compare(mod, inputs, params_np, rpc, exp_codegen_str)
-
-
-def get_relax_reshape_module(input_shape, output_shape, dtype):
+def get_relax_reshape_mod(input_shape, output_shape, dtype):
     """
     Args:
         input_shape (tuple): Input tensor shape
@@ -967,28 +588,7 @@ def get_relax_reshape_codegen(input_shape, output_shape, dtype):
     return expected_codegen_str
 
 
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "trials",
-    [
-        [(1, 3, 32, 32), (1, 4, -1, 32)],
-        [(1, 4, 8, 32), (1, 4, -1, 16)],
-        [(1, 64, 3, 3), (1, 32, 3, -1)],
-    ],
-)
-def test_reshape(dtype, trials, rpc):
-    low, high = -1, 1
-    (input_shape, output_shape) = trials
-    data = np.random.uniform(low, high, size=input_shape).astype(dtype)
-    inputs = [data]
-    mod = get_relax_reshape_module(input_shape, output_shape, dtype)
-    params_np = {}
-    expected_codegen = get_relax_reshape_codegen(input_shape, output_shape, dtype)
-    run_compare(mod, inputs, params_np, rpc, expected_codegen)
-
-
-def get_relax_global_avgpool_module(data_shape, keepdims, dtype):
+def get_relax_global_avgpool_mod(data_shape, keepdims, dtype):
     """
     Create a Relax module for Global Average Pooling (GAP).
 
@@ -1051,30 +651,7 @@ def get_global_avgpool_expected_codegen(input_shape, keep_dims, dtype):
     return exp_codegen
 
 
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "trials",
-    [
-        [(1, 64, 147, 147), True],
-        [(1, 256, 17, 17), False],
-        [(1, 1024, 14, 14), True],
-        [(1, 32, 256, 256), False],
-    ],
-)
-def test_global_avg_pool(dtype, trials, rpc):
-    """Test function for global average pooling."""
-    low, high = -1, 1
-    (input_shape, keep_dims) = trials
-    data = np.random.uniform(low, high, size=input_shape).astype(dtype)
-    inputs = [data]
-    mod = get_relax_global_avgpool_module(input_shape, keep_dims, dtype)
-    params_np = {}
-    exp_codegen_str = get_global_avgpool_expected_codegen(input_shape, keep_dims, dtype)
-    run_compare(mod, inputs, params_np, rpc, exp_codegen_str)
-
-
-def get_relax_global_maxpool_module(data_shape, keepdims, dtype):
+def get_relax_global_maxpool_mod(data_shape, keepdims, dtype):
     """
     Create a Relax module for Global Average Pooling (GAP).
 
@@ -1149,36 +726,3 @@ def get_global_maxpool_expected_codegen(input_shape, pool_size, stride, padding,
         },
     ]
     return exp_codegen
-
-
-@tvm.testing.requires_openclml
-@pytest.mark.parametrize("dtype", ["float32"])
-@pytest.mark.parametrize(
-    "trials",
-    [
-        [(1, 64, 147, 147), True],
-        [(1, 256, 17, 17), False],
-        [(1, 1024, 14, 14), True],
-        [(1, 32, 256, 256), False],
-    ],
-)
-def test_global_max_pool(dtype, trials, rpc):
-    """Test function for global average pooling."""
-    low, high = -1, 1
-    (input_shape, keep_dims) = trials
-    N, C, H, W = input_shape
-    pool_size = (H, W)
-    stride = (1, 1)
-    padding = (0, 0, 0, 0)
-    data = np.random.uniform(low, high, size=input_shape).astype(dtype)
-    inputs = [data]
-    mod = get_relax_global_maxpool_module(input_shape, keep_dims, dtype)
-    params_np = {}
-    exp_codegen_str = get_global_maxpool_expected_codegen(
-        input_shape, pool_size, stride, padding, "global_max", dtype
-    )
-    run_compare(mod, inputs, params_np, rpc, exp_codegen_str)
-
-
-if __name__ == "__main__":
-    tvm.testing.main()
